+++ trunk/NucleusLib/pnMessage/plMessage.cpp
@@ -7,7 +7,7 @@ plMessage::plMessage() {
 
 plMessage::~plMessage() {
     fSender->UnRef();
-    for (int i=0; i<fReceivers.getSize(); i++)
+    for (size_t i=0; i<fReceivers.getSize(); i++)
         fReceivers[i]->UnRef();
 }
 
@@ -17,7 +17,7 @@ void plMessage::IMsgRead(hsStream* S) {
     fSender = plResManager::inst->readKey(S);
     fSender->Ref();
     fReceivers.setSize(S->readInt());
-    for (int i=0; i<fReceivers.getSize(); i++) {
+    for (size_t i=0; i<fReceivers.getSize(); i++) {
         fReceivers[i] = plResManager::inst->readKey(S);
         fReceivers[i]->Ref();
     }
@@ -28,7 +28,7 @@ void plMessage::IMsgRead(hsStream* S) {
 void plMessage::IMsgWrite(hsStream* S) {
     plResManager::inst->writeKey(S, fSender);
     S->writeInt(fReceivers.getSize());
-    for (int i=0; i<fReceivers.getSize(); i++)
+    for (size_t i=0; i<fReceivers.getSize(); i++)
         plResManager::inst->writeKey(S, fReceivers[i]);
     S->writeDouble(fTimeStamp);
     S->writeInt(fBCastFlags);
